% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R, R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot FARS data}
\usage{
fars_map_state(state.num, year)

fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric code representing the state.}

\item{year}{A numeric value specifying the year for which the data should
be plotted.}
}
\value{
A plot of the places of accidents (as points) based on latitude and
 longitude on the map of the specified state.
 Returns a warning if invalid state code and/or an error in case an
 invalid year is specified.

A plot of the places of accidents (as points) based on latitude and
 longitude on the map of the specified state.
 Returns a warning if invalid state code and/or an error in case an
 invalid year is specified.
}
\description{
This functions plots the locations of incidents recorded on a map of
the specified state for the specified year.

This functions plots the locations of incidents recorded on a map of
the specified state for the specified year.
}
\examples{
\dontrun{
  # Map the accidents in the state of California in the year 2014.
  fars_map_state(6, 2014)
}

\dontrun{
  # Map the accidents in the state of California in the year 2014.
  fars_map_state(6, 2014)
}

}
