% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the name of the csv file to be read.}
}
\value{
Returns a tibble with the content of the specified file.
 It checks if the file specified exists and returns an error using
 the 'stop' function otherwise.
}
\description{
This function reads csv-files converts it into a tibble.
}
\examples{
\dontrun{
fars_read("accidents_2013.csv.bz2")
}

}
