% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of numeric values representing the years for which FARS
data should be read.}
}
\value{
A list of tibbles, each containing FARS data for a specific year
 with a 'MONTH' and a 'year' column.
 The function returns a warning if no FARS data set is available for
 the specified year.
}
\description{
This function reads data from the Fatality Analysis Reporting System (FARS) for
multiple years and returns a list of tibbles, each containing a 'MONTH' and
a 'year' column.
}
\details{
This function takes the user input uses the 'make_filename' function to
create the filenames to search for.
}
\examples{
\dontrun{
  # Read FARS data for years 2013 and 2014
  fars_data <- fars_read_years(c(2013, 2014))
}

}
