% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Sumamrize FARS data for multiple years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of numeric values representing the years for which FARS
data should be summarized.}
}
\value{
A single tibble containing the number of accidents for each month and
 year.
}
\description{
This functions reads data for the specified years then combines it into
a single data frame and summarizes the number of accidents for each month and
year.
}
\details{
This function utilizes the 'fars_read_years' function to obtain
 the data for the specified years.
}
\examples{
\dontrun{
  # Summarize total accidents for each month for the years 2013 and 2014
  fars_summarize_years(c(2013, 2014))
}

}
