% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create file name for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric or character vector specifying the year(s) of interest.}
}
\value{
Returns a character string representing the generated filename.
 The character string is formatted as: "accident_YEAR.csv.bz2".
}
\description{
This function generates a filename for FARS data for the specified year.
}
\examples{
\dontrun{
  make_filename(2013)
  make_filename(c(2013, 2014))
}

}
